+++ CMakeLists.txt
@@ -6,5 +6,5 @@ set(MATIO_MAT73 OFF CACHE BOOL OFF FORCE)
 add_subdirectory(deps/matio)
 include_directories(deps/matio/src build/deps/matio/src)
 
-add_executable(${PROJECT_NAME} src/main.c)
+add_executable(${PROJECT_NAME} src/main.c src/directivity.c)
 target_link_libraries(${PROJECT_NAME} matio)
